edit existing message, use embeds, minimal change to reactions

+++ modules/discord_connector.py
@@ -28,12 +28,35 @@ async def add_emoji_number_reactions(message, count):
     :param count: how many emojis to add
     :return: None
     """
+    
+    if count <= 0:
+        return
+        
+    # Only add reactions if necessary, 
+    cache_msg = await message.channel.fetch_message(message.id)
+    msg_emoji = [str(r.emoji) for r in cache_msg.reactions]
+    
+    emoji_to_remove = []
+    
+    for i,e in enumerate(msg_emoji):
+        if i >= count or i != vars.emoji_numbers.index(e):
+            emoji_to_remove.append(e)
+    
+    if len(emoji_to_remove) == len(msg_emoji):
+        await message.clear_reactions()
+        msg_emoji = []
+    else:
+        for e in emoji_to_remove:
+            await message.clear_reaction(e)
+            del(msg_emoji[msg_emoji.index(e)])
+    
     for i in range(0, count):
-        await message.add_reaction(vars.emoji_numbers[i])
+        if vars.emoji_numbers[i] not in msg_emoji:
+            await message.add_reaction(vars.emoji_numbers[i])
 
 
 class DiscordConnector:
-    def __init__(self, token, owner_id, refresh_time, tautulli_channel_id, tautulli_connector, analytics):
+    def __init__(self, token, owner_id, refresh_time, tautulli_channel_id, tautulli_connector, analytics, use_embeds, plex_pass):
         self.token = token
         self.owner_id = owner_id
         self.refresh_time = refresh_time
@@ -41,6 +64,8 @@ class DiscordConnector:
         self.tautulli_channel = None
         self.tautulli = tautulli_connector
         self.analytics = analytics
+        self.use_embeds = use_embeds
+        self.plex_pass = plex_pass
         self.client = discord.Client()
         self.on_ready = self.client.event(self.on_ready)
 
@@ -59,28 +84,74 @@ class DiscordConnector:
         :return: new discord.Message
         """
         new_message, count = self.tautulli.refresh_data()
-        await previous_message.delete()
-        new_message = await self.tautulli_channel.send(content=new_message)
-        await add_emoji_number_reactions(message=new_message, count=count)
-        bot_owner = self.client.get_user(self.owner_id)
-        reaction = ""
-        user = ""
+        
+        # For performance and aesthetics, edit the old message if 1) the old message is the newest message in the channel, or 2) if the only messages that are newer were written by this bot (which would be stream stop messages that have aleady been deleted)
+        use_old_message = False
+        async for msg in self.tautulli_channel.history(limit=100):
+            if msg.author != self.client.user:
+                use_old_message = False
+                break
+            elif msg.id == previous_message.id:
+                use_old_message = True
+                break
+        
+        if use_old_message:
+            if self.use_embeds:
+                await previous_message.edit(embed=new_message, content=None) # reset content to None to remove startup message
+            else:
+                await previous_message.edit(content=new_message, embed=None)
+            new_message = previous_message
+        else:
+            try:
+                await previous_message.delete()
+            except Exception as e:
+                info("Failed to delete old (specified) message: {}".format(e))
+                await tautulli_channel.purge(check=is_me)
+            if self.use_embeds:
+                new_message = await self.tautulli_channel.send(embed=new_message)
+            else:
+                new_message = await self.tautulli_channel.send(content=new_message)
+        
+        if self.plex_pass:
+            await add_emoji_number_reactions(message=new_message, count=count)
+            
+            cache_msg = await new_message.channel.fetch_message(new_message.id)
+            for reaction in cache_msg.reactions:
+                if reaction.count > 1:
+                    async for user in reaction.users():
+                        if user.id == self.owner_id:
+                            loc = vars.emoji_numbers.index(str(reaction.emoji))
+                            info("Stopping stream {}".format(loc))
+                            stopped_message = self.tautulli.stop_stream(stream_number=loc)
+                            info(stopped_message)
+                            end_notification = await self.tautulli_channel.send(content=stopped_message)
+                            await asyncio.sleep(min([5,self.refresh_time]))
+                            await end_notification.delete()
+                            await new_message.clear_reaction(str(reaction.emoji))
+                            return new_message
+        
+            reaction = ""
+            user = ""
 
-        def check(reaction, user):
-            return user == bot_owner
+            def check(reaction, user):
+                return user.id == self.owner_id and reaction.message.id == new_message.id and str(reaction.emoji) in vars.emoji_numbers
 
-        try:
-            reaction, user = await self.client.wait_for('reaction_add', timeout=float(self.refresh_time), check=check)
-        except asyncio.TimeoutError as e:
-            info(e)
-            pass
-        if reaction:
-            if user.id == self.owner_id:
+            try:
+                reaction, user = await self.client.wait_for('reaction_add', timeout=float(self.refresh_time), check=check)
+            except asyncio.TimeoutError as e:
+                info(e)
+                pass
+            else:
                 loc = vars.emoji_numbers.index(str(reaction.emoji))
+                info("Stopping stream {}".format(loc))
                 stopped_message = self.tautulli.stop_stream(stream_number=loc)
                 info(stopped_message)
                 end_notification = await self.tautulli_channel.send(content=stopped_message)
-                await end_notification.delete(delay=1.0)
+                await asyncio.sleep(min([5,self.refresh_time]))
+                await end_notification.delete()
+                await new_message.clear_reaction(str(reaction.emoji))
+        else:
+            await asyncio.sleep(self.refresh_time)
         return new_message
 
     async def get_tautulli_channel(self):
@@ -99,8 +170,9 @@ class DiscordConnector:
             async for msg in self.tautulli_channel.history(limit=100):
                 if msg.author == self.client.user:
                     last_bot_message_id = msg.id
+                    await msg.clear_reactions()
                     break
             if last_bot_message_id == "":
-                info("Couldn't find old message, sending 'Hello world!'")
-                await self.tautulli_channel.send(content="Hello world!")
+                info("Couldn't find old message, sending 'Start up'")
+                await self.tautulli_channel.send(content="Start up")
         return await self.tautulli_channel.fetch_message(last_bot_message_id)
